requirements.txt
setuptools
openai
fpdf
openpyxl
pillow